Move existing extensions to closer to actual class
